requirements.txt
Django==1.7
PyJWT==0.2.1
beautifulsoup4==4.3.2
facebook-sdk==0.4.0
oauthlib==0.6.3
psycopg2==2.5.4
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.7
requests==2.4.3
requests-oauthlib==0.4.1
six==1.8.0
stravalib==0.3.0
units==0.06
wsgiref==0.1.2
